feat(management): add the possibility to store others kinds of IdP
Closes gravitee-io/issues#3155

+++ src/components/identityProviders/identity-providers.component.ts
@@ -21,6 +21,7 @@ import PortalConfigService from '../../services/portalConfig.service';
 import { IScope } from 'angular';
 import OrganizationService from '../../services/organization.service';
 import EnvironmentService from '../../services/environment.service';
+import UserService from '../../services/user.service';
 import _ = require('lodash');
 
 const IdentityProvidersComponent: ng.IComponentOptions = {
@@ -38,6 +39,7 @@ const IdentityProvidersComponent: ng.IComponentOptions = {
     OrganizationService: OrganizationService,
     PortalConfigService: PortalConfigService,
     NotificationService: NotificationService,
+    UserService: UserService,
     $state: StateService,
     Constants,
     $rootScope: IScope
@@ -49,12 +51,16 @@ const IdentityProvidersComponent: ng.IComponentOptions = {
 
     this.$onInit = () => {
       this.identities.forEach((ipa: IdentityProviderActivation) => this.activatedIdps[ipa.identityProvider] = true);
+      this.socialIdentityProviders = _.filter(this.identityProviders, idp => idp.type === 'github' || idp.type === 'google' || idp.type === 'graviteeio_am' || idp.type === 'oidc');
+      this.internalIdentityProviders = _.filter(this.identityProviders, idp => idp.type === 'memory' || idp.type === 'gravitee' || idp.type === 'ldap');
     };
 
+    this.currentUserSource = UserService.currentUser.source;
     this.availableProviders = [
       { 'name': 'Gravitee.io AM', 'icon': 'perm_identity', 'type': 'graviteeio_am' },
       { 'name': 'Google', 'icon': 'google-plus', 'type': 'google' },
       { 'name': 'GitHub', 'icon': 'github-circle', 'type': 'github' },
+      { 'name': 'LDAP', 'icon': 'group', 'type': 'ldap' },
       { 'name': 'OpenID Connect', 'icon': 'perm_identity', 'type': 'oidc' }
     ];
 
@@ -63,25 +69,34 @@ const IdentityProvidersComponent: ng.IComponentOptions = {
     };
 
     this.delete = (provider: IdentityProvider) => {
-      let that = this;
-      $mdDialog.show({
-        controller: 'DialogConfirmController',
-        controllerAs: 'ctrl',
-        template: require('../dialog/confirmWarning.dialog.html'),
-        clickOutsideToClose: true,
-        locals: {
-          title: 'Are you sure you want to delete this identity provider?',
-          msg: '',
-          confirmButton: 'Delete'
-        }
-      }).then(function (response) {
-        if (response) {
-          IdentityProviderService.delete(provider).then(response => {
-            NotificationService.show('Identity provider \'' + provider.name + '\' has been deleted');
-            $state.go('management.settings.organization.identityproviders.list', {}, { reload: true });
-          });
-        }
-      });
+      if (this.currentUserSource === provider.id) {
+        var alert = $mdDialog.alert()
+          .title('Attention')
+          .textContent('Deleting this identity provider is forbidden since you have been authenticated with')
+          .ok('Close');
+
+        $mdDialog.show(alert);
+      } else {
+        let that = this;
+        $mdDialog.show({
+          controller: 'DialogConfirmController',
+          controllerAs: 'ctrl',
+          template: require('../dialog/confirmWarning.dialog.html'),
+          clickOutsideToClose: true,
+          locals: {
+            title: 'Are you sure you want to delete this identity provider?',
+            msg: '',
+            confirmButton: 'Delete'
+          }
+        }).then(function (response) {
+          if (response) {
+            IdentityProviderService.delete(provider).then(response => {
+              NotificationService.show('Identity provider \'' + provider.name + '\' has been deleted');
+              $state.go('management.settings.organization.identityproviders.list', {}, { reload: true });
+            });
+          }
+        });
+      }
     };
 
     this.saveForceLogin = () => {
@@ -124,6 +139,55 @@ const IdentityProvidersComponent: ng.IComponentOptions = {
           NotificationService.show(identityProviderId + ' is now ' + (this.activatedIdps[identityProviderId] ? 'enabled' : 'disabled'))
         );
       }
+    };
+
+    this.changeIdentityProviderStatus = (identityProvider: IdentityProvider) => {
+      if (this.currentUserSource === identityProvider.id) {
+        var alert = $mdDialog.alert()
+          .title('Attention')
+          .textContent('Disabling this identity provider is forbidden since you have been authenticated with')
+          .ok('Close');
+        identityProvider.enabled = true;
+        $mdDialog.show(alert);
+      } else {
+        IdentityProviderService.get(identityProvider.id).then(response => {
+          let idp: IdentityProvider = response;
+          idp.enabled = identityProvider.enabled;
+          IdentityProviderService.update(idp).then(response => {
+            NotificationService.show('Identity provider \'' + idp.name + '\' has been updated');
+          });
+        });
+      }
+    };
+
+    this.upward = (identityProvider: IdentityProvider) => {
+      IdentityProviderService.get(identityProvider.id).then(response => {
+        let idp: IdentityProvider = response;
+        idp.order = idp.order - 1;
+        IdentityProviderService.update(idp).then(response => {
+          NotificationService.show('Identity provider \'' + idp.name + '\' order has been changed with success');
+          this.refresh();
+        });
+      });
+    };
+
+    this.downward = (identityProvider: IdentityProvider) => {
+      IdentityProviderService.get(identityProvider.id).then(response => {
+        let idp: IdentityProvider = response;
+        idp.order = idp.order + 1;
+        IdentityProviderService.update(idp).then(response => {
+          NotificationService.show('Identity provider \'' + idp.name + '\' order has been changed with success');
+          this.refresh();
+        });
+      });
+    };
+
+    this.refresh = () => {
+      IdentityProviderService.list().then(response => {
+        this.identityProviders = response;
+        this.internalIdentityProviders = _.filter(this.identityProviders, idp => idp.type === 'memory' || idp.type === 'gravitee' || idp.type === 'ldap');
+
+      });
     };
   }
 };
